requirements.txt
langchain
langchain-openai
langchain-community
langchain-chroma
chromadb
python-dotenv
streamlit
openai
tiktoken
fastapi
uvicorn[standard]
